[Base] Split main.py into ui/ui_entry.py and core/daemon_entry.py
